# perf(core): Optimize `docusaurus start/serve`, fix `openBrowser()` perf issue on macOS