## Motivation

See https://github.com/jerrykingxyz/docusaurus/pull/1

The Rspack team helped us identify that the `openBrowser()` util we copied from Create-React-App is not optimized: 
https://github.com/facebook/create-react-app/blob/main/packages/react-dev-utils/openBrowser.js#L76

The Rspack compilation is delayed because it uses blocking `execSync` calls on macOS only. 

![image](https://github.com/user-attachments/assets/83ae0a8b-5315-48dc-9285-33be9961bc12)

![image](https://github.com/user-attachments/assets/d9983d4e-ed65-47e6-8994-eabdadeebedd)

In this env, the util tries to run an AppleScript so that we "reuse" an existing chromium tab using `https://localhost:3000` instead of opening a new one. 

```js
execSync('ps cax | grep "' + chromiumBrowser + '"');
execSync(
          'osascript openChrome.applescript "' +
            encodeURI(url) +
            '" "' +
            chromiumBrowser +
            '"',
          {
            cwd: __dirname,
            stdio: 'ignore',
          }
);
```

We'd like to keep this, since it's better DX, but not at the cost of slowing down the compiler. On my M3 Pro machine the compiler is delayed by 600ms on every `docusaurus start` or `docusaurus serve` command. 

The slowdown depends on the "browser position" in that array. Each browser to try takes ~150ms due to executing commands for each of them sequentially. Devs usually use "Google Chrome" so they have to try 4 browsers before the blocking shell calls stop (4 x 150ms ~= 600ms)

```js
const supportedChromiumBrowsers = [
      'Google Chrome Canary',
      'Google Chrome Dev',
      'Google Chrome Beta',
      'Google Chrome',
      'Microsoft Edge',
      'Brave Browser',
      'Vivaldi',
      'Chromium',
];
```





Moving to async non-blocking calls permits the bundler to start compiling while we are executing these commands. 

I also refactored the commands so that we only run one  `ps` shell command for all browsers to try, instead of running sequential `ps` commands one by one.


## Benchmark

### `docusaurus start`

For me (M3 / Google Chrome), 22% faster 🔥 

(measuring from CLI start to first compilation + browser opened) 


```bash
hyperfine --runs 5 'DOCUSAURUS_AB_BENCHMARK=true E2E_TEST=true yarn start:website' 'DOCUSAURUS_AB_BENCHMARK=false E2E_TEST=true yarn start:website'

Benchmark 1: DOCUSAURUS_AB_BENCHMARK=true E2E_TEST=true yarn start:website
  Time (mean ± σ):      4.121 s ±  0.177 s    [User: 6.925 s, System: 1.554 s]
  Range (min … max):    3.946 s …  4.377 s    5 runs
 
Benchmark 2: DOCUSAURUS_AB_BENCHMARK=false E2E_TEST=true yarn start:website
  Time (mean ± σ):      5.015 s ±  0.276 s    [User: 7.227 s, System: 1.791 s]
  Range (min … max):    4.652 s …  5.407 s    5 runs
 
Summary
  DOCUSAURUS_AB_BENCHMARK=true E2E_TEST=true yarn start:website ran
    1.22 ± 0.09 times faster than DOCUSAURUS_AB_BENCHMARK=false E2E_TEST=true yarn start:website
```

Note: this optimization will be even more important once we can fully leverage the Rspack persistent cache, since `openBrowser()` would become the main start bottleneck as we optimize other things.


### `docusaurus serve`

For me (M3 / Google Chrome), 30% faster 🔥 

```bash
hyperfine --runs 10 'DOCUSAURUS_AB_BENCHMARK=true yarn serve:website' 'DOCUSAURUS_AB_BENCHMARK=false yarn serve:website'

Benchmark 1: DOCUSAURUS_AB_BENCHMARK=true yarn serve:website
  Time (mean ± σ):      1.258 s ±  0.025 s    [User: 0.857 s, System: 0.183 s]
  Range (min … max):    1.229 s …  1.315 s    10 runs
 
Benchmark 2: DOCUSAURUS_AB_BENCHMARK=false yarn serve:website
  Time (mean ± σ):      1.621 s ±  0.019 s    [User: 0.891 s, System: 0.515 s]
  Range (min … max):    1.597 s …  1.657 s    10 runs
 
Summary
  DOCUSAURUS_AB_BENCHMARK=true yarn serve:website ran
    1.29 ± 0.03 times faster than DOCUSAURUS_AB_BENCHMARK=false yarn serve:website
```



## Test Plan

local benchmarks